requirements.txt
cryptography==3.2.1
python-telegram-bot==12.8
requests==2.25.1
